- bugfix: JDFParser was unable to parse files with relative filenames (e. g. ./../myJdf.jdf) - fixed Mac OS linking error in JDFLayout::moveElementsTo()JDFDateTimeEvaluation - JDFDateTimeState: keep compiler happy
git-svn-id: http://svn.cip4.org/svnroot/trunk@1590 25ee8a6d-831b-0410-907a-ff8528d65d49

+++ src/jdf/wrapper/JDFDateTimeEvaluation.cpp
@@ -281,7 +281,7 @@ namespace JDF{
 		
 		if (HasAttribute(atr_ValueDurationList))
 		{
-			JDFDuration duration=(range.GetRight().GetSecondsSince1970())-(range.GetLeft().GetSecondsSince1970());
+			JDFDuration duration=((double) range.GetRight().GetSecondsSince1970())-((double) range.GetLeft().GetSecondsSince1970());
 			return GetValueDurationList().IsPartOfRange(duration);
 		}
 		return true;
